requirements.txt
boto3
cryptography==2.9.2
pandas==1.1.2
numpy==1.18.4
toolz==0.9.0
jinja2==2.11.2
click==7.1.2
docker==4.2.0
apache-airflow-backport-providers-amazon==2020.6.24